+++ src/main/java/net/silentchaos512/mechanisms/block/generator/diesel/DieselGeneratorTileEntity.java
@@ -1,33 +1,35 @@
-package net.silentchaos512.mechanisms.block.generator.lava;
+package net.silentchaos512.mechanisms.block.generator.diesel;
 
 import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.fluid.Fluid;
 import net.minecraft.inventory.container.Container;
-import net.minecraft.item.BucketItem;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tags.FluidTags;
+import net.minecraft.tags.Tag;
 import net.minecraft.util.Direction;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraftforge.fluids.capability.IFluidHandler;
 import net.minecraftforge.fluids.capability.templates.FluidTank;
+import net.silentchaos512.mechanisms.api.IFluidContainer;
 import net.silentchaos512.mechanisms.block.generator.AbstractFluidFuelGeneratorTileEntity;
 import net.silentchaos512.mechanisms.init.ModTileEntities;
+import net.silentchaos512.mechanisms.util.InventoryUtils;
 import net.silentchaos512.mechanisms.util.TextUtil;
 
 import javax.annotation.Nullable;
 
-public class LavaGeneratorTileEntity extends AbstractFluidFuelGeneratorTileEntity {
+public class DieselGeneratorTileEntity extends AbstractFluidFuelGeneratorTileEntity {
     // Energy constants
     public static final int MAX_ENERGY = 50_000;
     public static final int MAX_SEND = 500;
-    public static final int ENERGY_CREATED_PER_TICK = 100;
-    public static final int TICKS_PER_MILLIBUCKET = 5;
+    public static final int ENERGY_CREATED_PER_TICK = 120;
+    public static final int TICKS_PER_MILLIBUCKET = 10;
 
-    public LavaGeneratorTileEntity() {
-        super(ModTileEntities.lavaGenerator, 2, MAX_ENERGY, 0, MAX_SEND, new FluidTank(4000, s -> s.getFluid().isIn(FluidTags.LAVA)));
-    }
+    static final Tag<Fluid> FUEL_TAG = new FluidTags.Wrapper(new ResourceLocation("forge", "diesel"));
 
-    public IFluidHandler getTank() {
-        return tank;
+    public DieselGeneratorTileEntity() {
+        super(ModTileEntities.dieselGenerator, 2, MAX_ENERGY, 0, MAX_SEND, new FluidTank(4000, s -> s.getFluid().isIn(FUEL_TAG)));
     }
 
     @Override
@@ -54,8 +56,8 @@ public class LavaGeneratorTileEntity extends AbstractFluidFuelGeneratorTileEntit
     @Override
     public boolean canInsertItem(int index, ItemStack stack, @Nullable Direction direction) {
         return index == 0
-                && stack.getItem() instanceof BucketItem
-                && ((BucketItem) stack.getItem()).getFluid().isIn(FluidTags.LAVA);
+                && InventoryUtils.isFilledFluidContainer(stack)
+                && tank.isFluidValid(IFluidContainer.getBucketOrContainerFluid(stack));
     }
 
     @Override
@@ -65,11 +67,11 @@ public class LavaGeneratorTileEntity extends AbstractFluidFuelGeneratorTileEntit
 
     @Override
     protected ITextComponent getDefaultName() {
-        return TextUtil.translate("container", "lava_generator");
+        return TextUtil.translate("container", "diesel_generator");
     }
 
     @Override
-    protected Container createMenu(int id, PlayerInventory playerInventory) {
-        return new LavaGeneratorContainer(id, playerInventory, this, this.fields);
+    protected Container createMenu(int id, PlayerInventory player) {
+        return new DieselGeneratorContainer(id, player, this, this.fields);
     }
 }