+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Diesel generator. Produces 120 FE/t, 10 ticks/mB (1.2M FE per bucket).
 ### Fixed
 - Added missing electrum recipes [#27]
 - Added missing textures for new metal blocks